+++ sudo_pair/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- The `-u` and `-g` flags can now both be used in exempt sessions.
+
 ## [0.11.1] - 2018-06-08
 
 ### Fixed