+++ sudo_pair/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - The `-u` and `-g` flags can now both be used in exempt sessions.
 
 ### Changed
+- Builds using Rust 2018
 - Errors are handled through the `failure` crate rather than `error-chain`.
 
 ## [0.11.1] - 2018-06-08